re-organized scss partials into separate directories
