+++ app/src/index.js
@@ -1,6 +1,6 @@
 import "innersvg-polyfill";
-import "./scss/shared.scss";
-import "./scss/homepage.scss";
+import "./scss/pages/shared.scss";
+import "./scss/pages/homepage.scss";
 import { translatePage } from "./utils/translate";
 
 translatePage();
