+++ app/src/infoPages.js
@@ -1,6 +1,6 @@
 // TODO: may not need this entry point, could keep if code splitting CSS
-import "./scss/shared.scss";
-import "./scss/info_pages.scss";
+import "./scss/pages/shared.scss";
+import "./scss/pages/info_pages.scss";
 import { translatePage } from "./utils/translate";
 
 translatePage();
