+++ app/src/scss/base/_base.scss
@@ -1,6 +1,4 @@
-@use "variables" as *;
-@use "mixins" as *;
-@import "~normalize.css";
+@use "../abstracts/" as *;
 
 // base styles for all pages
