+++ app/src/scss/components/_language_toggle.scss
@@ -1,5 +1,4 @@
-@use "variables" as *;
-@use "mixins" as *;
+@use "../abstracts/" as *;
 
 // NOTE: button layout & sizes differ somewhat between homepage & info pages
 // - info pages buttons are horizontally aligned
