+++ app/src/scss/components/_map.scss
@@ -1,5 +1,4 @@
-@use "variables" as *;
-@use "mixins" as *;
+@use "../abstracts/" as *;
 
 $map-width-default: 550px;
 $map-height-default: 48vh;
