+++ app/src/scss/components/_progress_circles.scss
@@ -1,5 +1,4 @@
-@use "variables" as *;
-@use "mixins" as *;
+@use "../abstracts/" as *;
 
 $circle-width: 24px;
 $show-min-width: calc($slides-max-width + $circle-width);
@@ -21,12 +20,12 @@ div#progress-indicator {
     width: $circle-width;
     height: $circle-width;
     margin-block-end: 0;
-    background-image: url("../../public/assets/png/oval_25_blank.png");
+    background-image: url("../../../public/assets/png/oval_25_blank.png");
     background-size: $circle-width;
     background-repeat: no-repeat;
 
     &.active {
-      background-image: url("../../public/assets/png/oval_25_filled.png");
+      background-image: url("../../../public/assets/png/oval_25_filled.png");
     }
   }
