+++ app/src/scss/components/_tr_orgs_modal.scss
@@ -1,5 +1,4 @@
-@use "variables" as *;
-@use "mixins" as *;
+@use "../abstracts/" as *;
 
 // tenants rights modal on homepage
 .tr-modal {
