--- app/src/scss/homepage.scss
@@ -1,9 +1,0 @@
-// styles used only by the homepage
-@use "slides";
-@use "map";
-@use "address_form";
-@use "search-validation-errors";
-@use "checking-sprite";
-@use "progress_circles";
-@use "light_box";
-@use "tr_orgs_modal";
