--- app/src/scss/info_pages.scss
@@ -1,4 +1,0 @@
-// styles used only by info pages
-@use "other_pages_main_content";
-@use "other_pages_nav";
-@use "embeds";
