+++ app/src/scss/layouts/_nav_main.scss
@@ -1,4 +1,4 @@
-@use "variables" as *;
+@use "../abstracts/" as *;
 
 // TODO: set up some spacing variables for UI consistency
 $nav-padding-inline: 32px;
