+++ app/src/scss/layouts/_other_pages_main_content.scss
@@ -1,5 +1,4 @@
-@use "variables" as *;
-@use "mixins" as *;
+@use "../abstracts/" as *;
 
 $sidebar-ideal-width: 250px;
