+++ app/src/scss/layouts/_other_pages_nav.scss
@@ -1,5 +1,4 @@
-@use "variables" as *;
-@use "mixins" as *;
+@use "../abstracts/" as *;
 
 nav.side-nav {
   h2,
