+++ app/src/scss/layouts/_slides.scss
@@ -1,5 +1,4 @@
-@use "variables" as *;
-@use "mixins" as *;
+@use "../abstracts/" as *;
 
 .slides {
   container-type: inline-size;
