+++ app/src/scss/pages/homepage.scss
@@ -1,0 +1,9 @@
+// styles used only by the homepage
+@use "../layouts/slides";
+@use "../components/map";
+@use "../components/address_form";
+@use "../components/search-validation-errors";
+@use "../components/checking-sprite";
+@use "../components/progress_circles";
+@use "../components/light_box";
+@use "../components/tr_orgs_modal";
