+++ app/src/scss/pages/info_pages.scss
@@ -1,0 +1,4 @@
+// styles used only by info pages
+@use "../layouts/other_pages_main_content";
+@use "../layouts/other_pages_nav";
+@use "../components/embeds";
