--- app/src/scss/shared.scss
@@ -1,6 +1,0 @@
-// styles shared between the homepage and info pages
-@use "base";
-@use "utility-classes";
-@use "nav_main";
-@use "buttons"; // used by primary navigation
-@use "language_toggle";
