+++ app/src/scss/utilities/_utility-classes.scss
@@ -1,5 +1,4 @@
-@use "variables" as *;
-@use "mixins" as *;
+@use "../abstracts/" as *;
 
 .hidden {
   display: none;
